filebench: fb_random64 cast UINT64_MAX to double to suppress warning
Apply a double cast to UINT64_MAX to suppress warning from Clang 18.1.6.

fb_random.c:75:39:
  error: implicit conversion from 'unsigned long' to 'double' changes value
  from 18446744073709551615 to 18446744073709551616
  [-Werror,-Wimplicit-const-int-float-conversion]
   75 |         random_normalized = (double)random / UINT64_MAX;
      |                                            ~ ^~~~~~~~~~
  /usr/include/stdint.h:119:23: note: expanded from macro 'UINT64_MAX'
  119 | # define UINT64_MAX             (__UINT64_C(18446744073709551615))
      |                                  ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
/usr/include/stdint.h:96:25: note: expanded from macro '__UINT64_C'
   96 | #  define __UINT64_C(c) c ## UL
      |                         ^~~~~~~
<scratch space>:18:1: note: expanded from here
   18 | 18446744073709551615UL
      | ^~~~~~~~~~~~~~~~~~~~~~

+++ fb_random.c
@@ -72,7 +72,7 @@ fb_random64(uint64_t *randp, uint64_t max, uint64_t round, avd_t avd)
 	 */
 	max = max - round;
 
-	random_normalized = (double)random / UINT64_MAX;
+	random_normalized = (double)random / (double)UINT64_MAX;
 	random = random_normalized * max;
 
 	if (round) {
